Refactor and add tests for determining a queue processor
