+++ tools/c7n_mailer/c7n_mailer/cli.py
@@ -5,13 +5,12 @@ import functools
 import logging
 from os import path
 
-import boto3
 import jsonschema
 import yaml
 from c7n_mailer import deploy, utils
 from c7n_mailer.azure_mailer import deploy as azure_deploy
 # from c7n_mailer.gcp_mailer import deploy as gcp_deploy
-from c7n_mailer.utils import get_provider, Providers
+from c7n_mailer.utils import session_factory, get_processor, get_provider, Providers
 
 AZURE_KV_SECRET_SCHEMA = {
     'type': 'object',
@@ -182,12 +181,6 @@ CONFIG_SCHEMA = {
 }
 
 
-def session_factory(mailer_config):
-    return boto3.Session(
-        region_name=mailer_config['region'],
-        profile_name=mailer_config.get('profile', None))
-
-
 def get_logger(debug=False):
     log_format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
     logging.basicConfig(level=logging.INFO, format=log_format)
@@ -272,19 +265,7 @@ def main():
         max_num_processes = args_dict.get('max_num_processes')
 
         # Select correct processor
-        # Import provider-specific processor modules lazily. Otherwise, all
-        # available providers must be installed to run or provision the
-        # mailer.
-        if provider == Providers.Azure:
-            from c7n_mailer.azure_mailer.azure_queue_processor import MailerAzureQueueProcessor
-            processor = MailerAzureQueueProcessor(mailer_config, logger)
-        elif provider == Providers.GCP:
-            from c7n_mailer.gcp_mailer.gcp_queue_processor import MailerGcpQueueProcessor
-            processor = MailerGcpQueueProcessor(mailer_config, logger)
-        elif provider == Providers.AWS:
-            from c7n_mailer.sqs_queue_processor import MailerSqsQueueProcessor
-            aws_session = session_factory(mailer_config)
-            processor = MailerSqsQueueProcessor(mailer_config, aws_session, logger)
+        processor = get_processor(provider, mailer_config, logger)
 
         # Execute
         if max_num_processes:
